+++ dev/experiments/crall/expt_drop8_2024-02-09.sh
@@ -2495,6 +2495,19 @@ initializer:
 "
 
 
+# shellcheck disable=SC2155
+export DVC_DATA_DPATH=$(geowatch_dvc --tags="phase3_data")
+# shellcheck disable=SC2155
+export DVC_EXPT_DPATH=$(geowatch_dvc --tags="phase3_expt")
+cd "$DVC_EXPT_DPATH"
+python -m geowatch.mlops.manager "status" --dataset_codes "Drop8-ARA-Median10GSD-V1"
+python -m geowatch.mlops.manager "list packages" --dataset_codes "Drop8-ARA-Median10GSD-V1"
+python -m geowatch.mlops.manager "list checkpoints" --dataset_codes "Drop8-ARA-Median10GSD-V1"
+python -m geowatch.mlops.manager "repackage checkpoints" --dataset_codes "Drop8-ARA-Median10GSD-V1"
+python -m geowatch.mlops.manager "gather packages" --dataset_codes "Drop8-ARA-Median10GSD-V1"
+python -m geowatch.mlops.manager "push packages" --dataset_codes "Drop8-ARA-Median10GSD-V1"
+
+
 # --------
 # ARA All Sensor AC FineTune, with Scotts rebalance, First Pass